remove the strangler + 💅
