+++ front-api/lib/api/sse/message_events.ts
@@ -1,0 +1,75 @@
+import { getConversationMessageType } from "@app/lib/api/assistant/conversation";
+import type { MessageStreamEvent } from "@app/lib/api/assistant/pubsub";
+import { getMessagesEvents } from "@app/lib/api/assistant/pubsub";
+import type { Authenticator } from "@app/lib/auth";
+import { ConversationResource } from "@app/lib/resources/conversation_resource";
+import { apiErrorForConversation } from "@front-api/lib/api/assistant/conversation/helper";
+import { streamEvents } from "@front-api/lib/api/sse/stream_events";
+import { apiError } from "@front-api/middlewares/utils";
+import type { Context } from "hono";
+import { z } from "zod";
+
+export const MessageParamSchema = z.object({
+  cId: z.string().min(1),
+  mId: z.string().min(1),
+});
+
+export type MessageEventsOptions = {
+  transformEvent: (auth: Authenticator, event: MessageStreamEvent) => unknown;
+};
+
+// Shared orchestration for both the v1 (public API) and private SSE
+// message-events routes; each supplies its own `transformEvent`. Public-API
+// stability rules ([BACK12]) apply to whatever the v1 caller emits.
+export async function streamMessageEventsForRoute(
+  ctx: Context,
+  auth: Authenticator,
+  {
+    conversationId,
+    messageId,
+    lastEventId,
+  }: { conversationId: string; messageId: string; lastEventId: string | null },
+  opts: MessageEventsOptions
+) {
+  const conversationRes =
+    await ConversationResource.fetchConversationWithoutContent(
+      auth,
+      conversationId
+    );
+  if (conversationRes.isErr()) {
+    return apiErrorForConversation(ctx, conversationRes.error);
+  }
+
+  const conversation = conversationRes.value;
+
+  const messageType = await getConversationMessageType(
+    auth,
+    conversation,
+    messageId
+  );
+  if (!messageType) {
+    return apiError(ctx, {
+      status_code: 404,
+      api_error: {
+        type: "message_not_found",
+        message: "The message you're trying to access was not found.",
+      },
+    });
+  }
+  if (messageType !== "agent_message") {
+    return apiError(ctx, {
+      status_code: 400,
+      api_error: {
+        type: "invalid_request_error",
+        message: "Events are only available for agent messages.",
+      },
+    });
+  }
+
+  return streamEvents({
+    ctx,
+    iterator: (signal) =>
+      getMessagesEvents(auth, { messageId, lastEventId, signal }),
+    transform: (event) => opts.transformEvent(auth, event),
+  });
+}
