+++ front-api/routes/sse/v1/w/[wId]/assistant/conversations/[cId]/messages/[mId]/events.test.ts
@@ -1,12 +1,11 @@
+import { fetchConversationMessages } from "@app/lib/api/assistant/messages";
 import type { MessageStreamEvent } from "@app/lib/api/assistant/pubsub";
 import { Authenticator } from "@app/lib/auth";
-import { MessageModel } from "@app/lib/models/agent/conversation";
 import { ConversationFactory } from "@app/tests/utils/ConversationFactory";
 import { createPublicApiMockRequest } from "@app/tests/utils/generic_public_api_tests";
 import { MembershipFactory } from "@app/tests/utils/MembershipFactory";
 import { UserFactory } from "@app/tests/utils/UserFactory";
 import { GLOBAL_AGENTS_SID } from "@app/types/assistant/assistant";
-import type { ModelId } from "@app/types/shared/model_id";
 import { honoApp } from "@front-api/app";
 import {
   asyncIteratorFrom,
@@ -16,22 +15,25 @@ import {
 } from "@front-api/tests/utils/sse";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
-async function getMessageByRank(
+async function getMessageSIdByRank(
   auth: Authenticator,
-  conversationId: ModelId,
+  conversationId: string,
   rank: number
-): Promise<MessageModel> {
-  const message = await MessageModel.findOne({
-    where: {
-      workspaceId: auth.getNonNullableWorkspace().id,
-      conversationId,
-      rank,
-    },
+): Promise<string> {
+  const messagesRes = await fetchConversationMessages(auth, {
+    conversationId,
+    limit: 100,
+    lastRank: null,
+    viewType: "light",
   });
+  if (messagesRes.isErr()) {
+    throw messagesRes.error;
+  }
+  const message = messagesRes.value.messages.find((m) => m.rank === rank);
   if (!message) {
     throw new Error(`No message found at rank ${rank}`);
   }
-  return message;
+  return message.sId;
 }
 
 vi.mock("@app/lib/api/assistant/pubsub", async (importOriginal) => {
@@ -84,9 +86,17 @@ async function setupAgentMessage() {
     agentConfigurationId: GLOBAL_AGENTS_SID.DUST,
     messagesCreatedAt: [new Date()],
   });
-  const agentMessage = await getMessageByRank(userAuth, conversation.id, 1);
-  const userMessage = await getMessageByRank(userAuth, conversation.id, 0);
-  return { workspace, key, conversation, agentMessage, userMessage };
+  const agentMessageSId = await getMessageSIdByRank(
+    userAuth,
+    conversation.sId,
+    1
+  );
+  const userMessageSId = await getMessageSIdByRank(
+    userAuth,
+    conversation.sId,
+    0
+  );
+  return { workspace, key, conversation, agentMessageSId, userMessageSId };
 }
 
 describe("GET /api/sse/v1/w/[wId]/assistant/conversations/[cId]/messages/[mId]/events", () => {
@@ -124,13 +134,13 @@ describe("GET /api/sse/v1/w/[wId]/assistant/conversations/[cId]/messages/[mId]/e
   });
 
   it("returns 400 when the target message is not an agent message", async () => {
-    const { workspace, key, conversation, userMessage } =
+    const { workspace, key, conversation, userMessageSId } =
       await setupAgentMessage();
 
     const response = await getMessageEvents(
       workspace.sId,
       conversation.sId,
-      userMessage.sId,
+      userMessageSId,
       key.secret
     );
 
@@ -141,7 +151,7 @@ describe("GET /api/sse/v1/w/[wId]/assistant/conversations/[cId]/messages/[mId]/e
   });
 
   it("streams an empty SSE response when no events are produced", async () => {
-    const { workspace, key, conversation, agentMessage } =
+    const { workspace, key, conversation, agentMessageSId } =
       await setupAgentMessage();
 
     vi.mocked(getMessagesEvents).mockImplementation(emptyAsyncIterator);
@@ -149,7 +159,7 @@ describe("GET /api/sse/v1/w/[wId]/assistant/conversations/[cId]/messages/[mId]/e
     const response = await getMessageEvents(
       workspace.sId,
       conversation.sId,
-      agentMessage.sId,
+      agentMessageSId,
       key.secret
     );
 
@@ -157,7 +167,7 @@ describe("GET /api/sse/v1/w/[wId]/assistant/conversations/[cId]/messages/[mId]/e
   });
 
   it("streams events for an agent message to the client", async () => {
-    const { workspace, key, conversation, agentMessage } =
+    const { workspace, key, conversation, agentMessageSId } =
       await setupAgentMessage();
 
     vi.mocked(getMessagesEvents).mockImplementation(
@@ -167,7 +177,7 @@ describe("GET /api/sse/v1/w/[wId]/assistant/conversations/[cId]/messages/[mId]/e
     const response = await getMessageEvents(
       workspace.sId,
       conversation.sId,
-      agentMessage.sId,
+      agentMessageSId,
       key.secret
     );
