+++ front-api/routes/sse/v1/w/[wId]/index.ts
@@ -14,10 +14,7 @@ const app = publicApiApp();
 app.use("*", publicApiAuth);
 
 app.route("/assistant/conversations/:cId/events", conversationEvents);
-app.route(
-  "/assistant/conversations/:cId/messages/:mId/events",
-  messageEvents
-);
+app.route("/assistant/conversations/:cId/messages/:mId/events", messageEvents);
 app.route("/mcp/requests", mcpRequests);
 
 export default app;
