+++ front-api/routes/sse/w/[wId]/assistant/conversations/[cId]/messages/[mId]/events.ts
@@ -1,15 +1,15 @@
+import type { MessageEventsOptions } from "@front-api/lib/api/sse/message_events";
+import {
+  MessageParamSchema,
+  streamMessageEventsForRoute,
+} from "@front-api/lib/api/sse/message_events";
 import { SseQuerySchema } from "@front-api/lib/api/sse/stream_events";
 import { workspaceApp } from "@front-api/middlewares/ctx";
 import { streamingTag } from "@front-api/middlewares/streaming";
 import { validate } from "@front-api/middlewares/validator";
-import type { MessageEventsOptions } from "@front-api/routes/sse/v1/w/[wId]/assistant/conversations/[cId]/messages/[mId]/events";
-import {
-  MessageParamSchema,
-  streamMessageEventsForRoute,
-} from "@front-api/routes/sse/v1/w/[wId]/assistant/conversations/[cId]/messages/[mId]/events";
 
 // Mounted at /api/sse/w/:wId/assistant/conversations/:cId/messages/:mId/events.
-// Handler logic lives in the v1 sibling file.
+// Handler logic lives in `@front-api/lib/api/sse/message_events`.
 
 const PRIVATE_OPTIONS: MessageEventsOptions = {
   transformEvent: (_auth, event) => event,
